Adjusted use of tangokans to strongly prefer new Tangomon.
For Tangomon you already have, the game now only opts to use a
tangokan 10% of the time, meaning you have on average 10 attempts
to encounter the Tangomon you wish to catch before you start getting
duplicates. The reason I didn't eliminate it entirely is because
I don't want it to be too perfect, and besides, some redundancy
protects the player's safety from defeat and failing tests.