Back up save files before saving them.
I actually stopped using Tangomon when I lost *everything*. All of
my vocabulary, gone in an instant, because the game crashed at just
the right time for save_slots.json to get emptied out, but not
replaced. Now, it always saves a backup prior to saving, then deletes
the backup once saving is finished. If a backup file is detected on
startup, that backup file is used instead of the regular save file.