Removed the time limit for tangoji entry.
It occurs to me that this time limit, while it makes the game harder,
also makes Tangomon less efficient as a learning tool. The best way
to remember something is to go over for it for a really long time and
then finally remember it on your own. And if you simply don't have a
clue, you can always just press "Enter" and forfeit the tangoji.